% Generated by roxygen2 (4.0.0): do not edit by hand
\name{BM}
\alias{BM}
\title{Bateman-mukai estimators for MA data}
\usage{
BM(delta_m, delta_v)
}
\arguments{
\item{delta_m}{numeric, change in mean fitness}

\item{delta_v}{numeric, change in among-line variance}
}
\value{
U numeric BM estimator of mutation rate

Ea numeric BM estimator of effect-size
}
\description{
Bateman-mukai estimators for MA data
}
\examples{
BM(-0.01, 0.001)
x <- rma_normal(n=100, s=0.01, Vs=1e-9, Ve=1e-3, Ut=10)
BM(mean(x), var(x) - 1e-3)
}

