% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dma_gamma}
\alias{dma_gamma}
\title{Density function of gamma dfe}
\usage{
dma_gamma(obs, shape, rate, Ve, Ut, log = TRUE)
}
\arguments{
\item{obs,}{numeric, observed fitnesses}

\item{shape,}{numeric, shape of gamma distribution}

\item{rate,}{numeric, rate of gamma distribution}

\item{Ve,}{numeric, experimental variance}

\item{Ut,}{numeric, mutation rate}

\item{log,}{boolean, return log-liklihood (default=TRUE)}
}
\description{
Density function of gamma dfe
}

