% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamma.r
\name{dma_gamma_r}
\alias{dma_gamma_r}
\title{Calculate probability density of a set of fitness measures under a
Normal-Gamma convolution model}
\usage{
dma_gamma_r(w, a, B, Ve, Ut, log = FALSE)
}
\arguments{
\item{w}{numeric, observed data}

\item{a}{numeric,  shape parameter for Gamma}

\item{B}{numeric, Scale paramater for Gamma}

\item{Ve}{numeric, envrionmental variance}

\item{Ut}{numeric, expected number of mutations over the length of the}

\item{log}{boolean, return log liklihood}
}
\description{
Calculate probability density of a set of fitness measures under a
Normal-Gamma convolution model
}
\examples{
dma_gamma( c(0.1,0.01), 0.05, 2, 1e-4, 2)
}

