% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dma_IG}
\alias{dma_IG}
\title{Density function for inverse-gaussian dfe}
\usage{
dma_IG(obs, mean, shape, Ve, Ut, log = TRUE)
}
\arguments{
\item{obs}{numeric, observed fitnesses}

\item{mean}{numeric, mean effect-size of mutations}

\item{shape}{numeric, shape of underlying DFE}

\item{Ve}{numeric variance of experimental system}

\item{Ut}{numeric mutation rate}

\item{log}{boolean return log liklihood (default=TRUE)}
}
\description{
Density function for inverse-gaussian dfe
}

