% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dma_IG_known}
\alias{dma_IG_known}
\title{Density function of gamma dfe with known mutation counts}
\usage{
dma_IG_known(obs, mean, shape, Ve, k, p_neutral, log = TRUE)
}
\arguments{
\item{obs, }{numeric, observed fitnesses}

\item{mean, }{numeric, rate of inverse gaussian DFE}

\item{shape, }{numeric, shape of inverse gaussian DFE}

\item{Ve, }{numeric, experimental variance}

\item{k, }{numeric, vector of knonw mutation counts}

\item{p_neutral}{numeric,proportion of all mutations that have no effect}

\item{log}{logical return log-liklihood (defaults to true)}
}
\value{
numeric (log-) liklihood of the specfified model and data
}
\description{
Density function of gamma dfe with known mutation counts
}
\examples{
set.seed(123)
mu <- rpois(50, 10)
w <- rma_known_IG(shape=1, mean=0.1, Ve=0.01, k=mu,p_neutral=0.7) 
dma_IG_known(w, shape=1, mean=0.1, Ve=0.01, k=mu, p_neutral=0.75)
dma_IG_known(w, shape=1, mean=0.1, Ve=0.01, k=mu, p_neutral=0.65)
}

