% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dma_normal}
\alias{dma_normal}
\title{Calculate log-likilhood for a single observation given}
\usage{
dma_normal(obs, s, Vs, Ve, Ut, log = FALSE)
}
\arguments{
\item{obs}{numeric, a single observed fitness from a an MA lne}

\item{s}{numeric,  mean fitness of mutations}

\item{Vs}{numeric, variance in fitness of mutations}

\item{Ve}{numeric, envrionmental variance}

\item{Ut}{numeric, expected number of mutations over the length of the
experiment}

\item{log}{logical, return log-transformed likilihod?}
}
\value{
numeric, likilood of observed data given paramaters
}
\description{
This function calculate the liklihood of a given observed fitness under
a model in which the fitness effects of mutations are normall distributed
}
\examples{
Changinh mutation rates:
dma_normal(0.1, 0.1, 0.01, 0.01, 1)
dma_normal(0.1, 0.1, 0.01, 0.01, 3)
}

