% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dma_normal_known}
\alias{dma_normal_known}
\title{Density function for normal dfe with known mutations}
\usage{
dma_normal_known(obs, a, Va, Ve, k, p_neutral, log = FALSE)
}
\arguments{
\item{obs}{numeric, observed fitnesses}

\item{a}{numeric, mean effect-size of mutations}

\item{Va}{numeric variance of dfe}

\item{Ve}{numeric variance of experimental system}

\item{k}{integer, vector of mutation counts}

\item{log}{boolean return log liklihood (default=TRUE)}

\item{p_nuetral, }{global proportion of neutral mutations}
}
\description{
Density function for normal dfe with known mutations
}

