% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fgm.r
\name{estimate_FGM_moments}
\alias{estimate_FGM_moments}
\title{Estimate moments of d.f.e. as defined by fishers geometic model}
\usage{
estimate_FGM_moments(n = 1e+05, start, Vm, FUN = sq_dist, plot = FALSE)
}
\arguments{
\item{n}{numeric number of observations from which to estimate the model}

\item{start,}{numeric vector of starting values}

\item{Vm,}{variance of mutation-size}

\item{FUN,}{function, fitness funciton}

\item{plot,}{boolean, make density plot of dfe (default=FALSE)}
}
\description{
Estimate moments of d.f.e. as defined by fishers geometic model
}

