% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gamma.r
\name{fit_ma_gamma}
\alias{fit_ma_gamma}
\title{Calculate probability density of a set of fitness measures under a
Normal-Gamma convolution model}
\usage{
fit_ma_gamma(obs, fixed = list(), start = list(), verbose = FALSE)
}
\arguments{
\item{fixed}{a named list containing values for any of the above paramaters,
that should be fixed during the maximum likelihood estimation}

\item{start}{a named list contaning starting values for any of the above
paramaters
Note, all paramaters must be given either a fixed or a starting value}

\item{a}{numeric,  shape parameter for Gamma}

\item{B}{numeric, Scale paramater for Gamma}

\item{Ve}{numeric, envrionmental variance}

\item{Ut}{numeric, expected number of mutations over the length of the}

\item{a}{numeric,  shape parameter for Gamma}

\item{B}{numeric, Scale paramater for Gamma}

\item{Ve}{numeric, envrionmental variance}

\item{Ut}{numeric, expected number of mutations over the length of the}
}
\description{
Calculate probability density of a set of fitness measures under a
Normal-Gamma convolution model
}
\examples{
dma_gamma( c(0.1,0.01), 0.05, 2, 1e-4, 2)
}

