% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_normal}
\alias{grad_normal}
\title{Gradient function of the normal model}
\usage{
grad_normal(obs, a, Va, Ve, Ut)
}
\arguments{
\item{obs}{numeric, observed fitnesses}

\item{a}{numeric, mean effect-size of mutations}

\item{Va}{numeric variance of dfe}

\item{Ve}{numeric variance of experimental system}

\item{Ut}{numeric mutation rate}
}
\description{
Gradient function of the normal model
}

