% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe-package.r
\docType{data}
\name{ma_control}
\alias{ma_control}
\title{Fitness data for the control lines of a mutation accumulation experiment}
\format{a vector of floating point fitness values}
\usage{
ma_control
}
\description{
These data represent the mean fitness of the ancestral lines used in a
mutation accumulation. The most common use of this data will be to estimate
the 'experimental variance' (Ve) of the the fitness assay
\code{var(ma_control)}
}
\keyword{datasets}

