% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfe-package.r
\docType{data}
\name{ma_experimental}
\alias{ma_experimental}
\title{Fitness data for a set of mutation accumulation lines}
\format{a vector of floating point fitness values}
\usage{
ma_experimental
}
\description{
These data represent the mean fitness of a set of mutation accumulation 
lines. These data can be used to calculate the change in fitness of these i
lines relative to  mean fitness of the control data: \code{ma_experimental -
mean(ma_control)}
}
\keyword{datasets}

