% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gamma.r
\name{mom_ma_gamma}
\alias{mom_ma_gamma}
\title{M.O.M estimate of gamma model with known mutation rate}
\usage{
mom_ma_gamma(obs, Ut, Ve)
}
\arguments{
\item{obs,}{numeric, vector of observed fitnesses}

\item{Ut,}{numeric, mutation rate}

\item{Ve,}{numeric, experimental variance}
}
\description{
M.O.M estimate of gamma model with known mutation rate
}

