% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normal.r
\name{mom_ma_normal}
\alias{mom_ma_normal}
\title{Method of moments estimators for the normally distributed DFE}
\usage{
mom_ma_normal(obs, Ve)
}
\arguments{
\item{obs}{observed fitness values}

\item{Ve}{estimate of btween-line (experimental) variance}
}
\value{
a Mean mutational effect size

Va Variance of d.f.e

Ut expected number of mutaitons over the experiment
}
\description{
Method of moments estimators for the normally distributed DFE
}
\examples{
set.seed(123)
w <- rma_normal(n=20, a=0.1, Va=0.01, Ve=0.01,Ut=1)
mom_ma_normal(w, 0.01)
}

