% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rma_FGM}
\alias{rma_FGM}
\title{Simulate fitness effects under fishers geometric model}
\usage{
rma_FGM(n, start, Vm, Ve, Ut, FUN = sq_dist)
}
\arguments{
\item{n}{numeric, number of lines to simulate}

\item{start}{numeric,  vector describing the starting position relative to an
optimum at teh origin (can take any number of dimensions)}

\item{Vm}{numeric, variance in size of mutations}

\item{Ve}{numeric, envrionmental variance}

\item{Ut}{numeric, expected number of mutations over the length of the
experiment}
}
\value{
numeric, a vector of fitnesses
}
\description{
This function simulates fitness effects under a model in which the fitness
distribution is simulated as coming from Fisher's geometic model
}
\examples{
set.seed(123)
w <-
}

