% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rma_gamma}
\alias{rma_gamma}
\title{Simulate fitness effects under of Gamma model}
\usage{
rma_gamma(n, s, B, Ve, Ut)
}
\arguments{
\item{n}{numeric, number of lines to simulate}

\item{s}{numeric,  shape parameter for Gamma}

\item{B}{numeric, Scale paramater for Gamma}

\item{Ve}{numeric, envrionmental variance}

\item{Ut}{numeric, expected number of mutations over the length of the
experiment}
}
\value{
numeric, a vector of fitnesses
}
\description{
This function simulates fitness effects under a model in which the fitness
distribution of mutations takes a Gamma distribution
}
\examples{
set.seed(123)
w <- rma_normal(20, 0.1, 1 0.01, 1)
}

