% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.r
\name{rma_known_normal}
\alias{rma_known_normal}
\title{Simulate fitness effects ith a known number of mutations per line and a Normal DFE}
\usage{
rma_known_normal(a, Va, Ve, k, p_neutral)
}
\arguments{
\item{a}{numeric,  mean fitness of mutations}

\item{Va}{numeric, variance in fitness of mutations}

\item{Ve}{numeric, envrionmental variance}

\item{k}{integer, total number of mutations for each line}

\item{p_neutral}{numeric, (global) proportion of those mutation with no fitness effect}
}
\value{
numeric, a vector of fitnesses
}
\description{
This function simulates fitness effects under a model in which the fitness
distribution of mutations is normally distributed
}
\examples{
set.seed(123)
k <- rpois(20,9)
w <- rma_normal(a=0.1, Va=0.01, Ve=0.01, k=k, p_neutral=0.7)
mean(w)
}

