% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rma_normal}
\alias{rma_normal}
\title{Simulate fitness effects under of normal model}
\usage{
rma_normal(n, s, Vs, Ve, Ut)
}
\arguments{
\item{n}{numeric, number of lines to simulate}

\item{s}{numeric,  mean fitness of mutations}

\item{Vs}{numeric, variance in fitness of mutations}

\item{Ve}{numeric, envrionmental variance}

\item{Ut}{numeric, expected number of mutations over the length of the
experiment}
}
\value{
numeric, a vector of fitnesses
}
\description{
This function simulates fitness effects under a model in which the fitness
distribution of mutations is normally distributed
}
\examples{
set.seed(123)
w <- rma_normal(20, 0.1, 0.01, 0.01, 1)
dma_normal(w, 0.1, 0.01,0.01, 1)
dma_normal(w, 0.1, 0.01,0.01, 3)
}

