% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dma_normal}
\alias{dma_normal}
\title{Density function for normal dfe}
\usage{
dma_normal(obs, a, Va, Ve, Ut, log = TRUE)
}
\arguments{
\item{obs}{numeric, observed fitnesses}

\item{a}{numeric, mean effect-size of mutations}

\item{Va}{numeric variance of dfe}

\item{Ve}{numeric variance of experimental system}

\item{Ut}{numeric mutation rate}

\item{log}{boolean return log liklihood (default=TRUE)}
}
\description{
Density function for normal dfe
}

