% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gamma.r
\name{fit_ma_gamma}
\alias{fit_ma_gamma}
\title{Fit a MA-model with Gamma dfe}
\usage{
fit_ma_gamma(obs, fixed = list(), start = list(), verbose = FALSE)
}
\arguments{
\item{fixed}{a named list containing values for any of the above paramaters,
that should be fixed during the maximum likelihood estimation}

\item{start}{a named list contaning starting values for any of the above
paramaters
Note, all paramaters must be given either a fixed or a starting value}

\item{a}{numeric,  shape parameter for Gamma}

\item{B}{numeric, Scale paramater for Gamma}

\item{Ve}{numeric, envrionmental variance}

\item{Ut}{numeric, expected number of mutations over the length of the}
}
\description{
Fit a MA-model with Gamma dfe
}

