% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normal.r
\name{fit_ma_normal}
\alias{fit_ma_normal}
\title{Find the maximum liklihood estimate of paramaters in MA model}
\usage{
fit_ma_normal(obs, fixed = NULL, starts = NULL, verbose = TRUE)
}
\arguments{
\item{obs}{observed fitness values}

\item{fixed}{named list of model paramater values to fix}

\item{starts}{named list of starting values of varying parameters}

\item{verbose}{logical, be verbose}
}
\value{
mle fit object
}
\description{
Find the maximum liklihood estimate of paramaters in MA model
}
\examples{
set.seed(123)
w <- rma_normal(20, 0.1, 0.01, 0.01, 1)
fit_ma_normal(w, fixed=list(Ve=0.01), starts=list(s=0.01))
}

