% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gamma.r
\name{moments_gamma}
\alias{moments_gamma}
\title{Get mean and variance of a Gamma distribution given shape and rate
paramaters}
\usage{
moments_gamma(shape, rate)
}
\arguments{
\item{shape}{numeric,  shape parameter for Gamma}

\item{rate}{numeric, Scale paramater for Gamma}
}
\description{
Get mean and variance of a Gamma distribution given shape and rate
paramaters
}

