% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fgm.r
\name{rma_FGM}
\alias{rma_FGM}
\title{Simulate fitness effects under fishers geometric model}
\usage{
rma_FGM(n, start, Vm, Ve, Ut, FUN = gaussian_fitness)
}
\arguments{
\item{n}{numeric, number of lines to simulate}

\item{start}{numeric,  vector describing the starting position relative to an
optimum at teh origin (can take any number of dimensions)}

\item{Vm}{numeric, variance in size of mutations}

\item{Ve}{numeric, envrionmental variance}

\item{Ut}{numeric, expected number of mutations over the length of the
experiment}

\item{FUN,}{function, fitness funciton}
}
\value{
numeric, a vector of fitnesses
}
\description{
This function simulates fitness effects under a model in which the fitness
distribution is simulated as coming from Fisher's geometic model
}
\examples{
set.seed(123)
w <- rma_FGM(100, start=c(0.04,-0.04), Ve=1e-4, Vm=0.003, Ut=2)
mean(w)
}

