% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gamma.r
\name{rma_gamma}
\alias{rma_gamma}
\title{Simulate fitness effects under of Gamma model}
\usage{
rma_gamma(n, a, B, Ve, Ut)
}
\arguments{
\item{n}{numeric, number of lines to simulate}

\item{a}{numeric,  shape parameter for Gamma}

\item{B}{numeric, Scale paramater for Gamma}

\item{Ve}{numeric, envrionmental variance}

\item{Ut}{numeric, expected number of mutations over the length of the
experiment}
}
\value{
numeric, a vector of fitnesses
}
\description{
This function simulates fitness effects under a model in which the fitness
distribution of mutations takes a Gamma distribution
}
\examples{
set.seed(123)
w <- rma_gamma(20, 0.1, 2, 1e-4, 3)
dma_gamma(w, 0.1, 2, 1e-4, 3)
}

